requirements.txt
discord.py==1.7.2
python-dotenv==0.17.1
asyncio==3.4.3